services/api/python/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI/ML
openai==1.10.0
anthropic==0.25.0
numpy==1.24.3
scikit-learn==1.3.2

# Data processing
pandas==2.1.4
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
